requirements.txt
scipy
tqdm
torch
matplotlib
numpy
torchvision
Pillow
